+++ package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package>
-  <name>dynamixel_control_hw</name>
+  <name>dynamixel_control</name>
   <version>0.1.0</version>
   <description>An interface to the Dynamixel actuators for ROS control</description>
